Refactor SubCommand class to inherit from CommandAttrs and streamline initialization logic
